Disable the ObjectStore Realm caching
Might fix https://secure.helpscout.net/conversation/343482861/0/?folderId=366246.

Since we still have the cache of obj-c RLMRealm instances the
ObjectStore cache is at best a harmless waste of time, but if the Realm
instance outlives the owning RLMRealm (perhaps due to a collection
notifier for that Realm being in the middle of running when the RLMRealm
is deallocated?) the existing Realm instance may be handed to a new
RLMRealm, which is not handled by the existing code and may cause more
odd problems.